requirements.txt
azure_communication_email==1.0.0
flask==3.0.1
python-dotenv==1.0.1